+++ src/features/tracing/nrfml.ts
@@ -18,6 +18,7 @@ import { fileExtension, sinkName, TraceFormat } from './traceFormat';
 import {
     getManualDbFilePath,
     getSerialPort,
+    setDetectingTraceDb,
     setTaskId,
     setTracePath,
     setTraceSize,
@@ -86,6 +87,31 @@ const sinkConfig = (format: TraceFormat, filePath: string, device?: Device) =>
         },
     } as const);
 
+function detectTraceDB(progress: nrfml.Progress, detectedTraceDB: unknown) {
+    if (
+        progress.meta?.modem_db_path != null &&
+        detectedTraceDB !== progress.meta?.modem_db_path
+    ) {
+        detectedTraceDB = progress.meta?.modem_db_path;
+        logger.info(`Using trace DB ${detectedTraceDB}`);
+    }
+    return detectedTraceDB;
+}
+
+function detectModemFwUuid(
+    progress: nrfml.Progress,
+    detectedModemFwUuid: unknown
+) {
+    if (
+        progress.meta?.modem_db_uuid != null &&
+        detectedModemFwUuid !== progress.meta?.modem_db_uuid
+    ) {
+        detectedModemFwUuid = progress.meta?.modem_db_uuid;
+        logger.info(`Detected modem firmware with UUID ${detectedModemFwUuid}`);
+    }
+    return detectedModemFwUuid;
+}
+
 const convertTraceFile =
     (sourcePath: string): TAction =>
     (dispatch, getState) => {
@@ -120,22 +146,13 @@ const convertTraceFile =
                 dispatch(setTaskId(null));
             },
             progress => {
-                if (
-                    progress.meta?.modem_db_uuid != null &&
-                    detectedModemFwUuid !== progress.meta?.modem_db_uuid
-                ) {
-                    detectedModemFwUuid = progress.meta?.modem_db_uuid;
-                    logger.info(
-                        `Detected modem firmware with UUID ${detectedModemFwUuid}`
+                if (!manualDbFilePath) {
+                    detectedModemFwUuid = detectModemFwUuid(
+                        progress,
+                        detectedModemFwUuid
                     );
-                }
 
-                if (
-                    progress.meta?.modem_db_path != null &&
-                    detectedTraceDB !== progress.meta?.modem_db_path
-                ) {
-                    detectedTraceDB = progress.meta?.modem_db_path;
-                    logger.info(`Using trace DB ${detectedTraceDB}`);
+                    detectedTraceDB = detectTraceDB(progress, detectedTraceDB);
                 }
 
                 progress.data_offsets
@@ -165,6 +182,9 @@ const startTrace =
         const filePath =
             path.join(getAppDataDir(), filename) + fileExtension(traceFormat);
         const manualDbFilePath = getManualDbFilePath(getState());
+        if (!manualDbFilePath && traceFormat === 'pcap') {
+            dispatch(setDetectingTraceDb(true));
+        }
 
         let detectedModemFwUuid: unknown = '';
         let detectedTraceDB: unknown = '';
@@ -194,22 +214,14 @@ const startTrace =
                 }
             },
             progress => {
-                if (
-                    progress.meta?.modem_db_uuid != null &&
-                    detectedModemFwUuid !== progress.meta?.modem_db_uuid
-                ) {
-                    detectedModemFwUuid = progress.meta?.modem_db_uuid;
-                    logger.info(
-                        `Detected modem firmware with UUID ${detectedModemFwUuid}`
+                if (!manualDbFilePath) {
+                    detectedModemFwUuid = detectModemFwUuid(
+                        progress,
+                        detectedModemFwUuid
                     );
-                }
 
-                if (
-                    progress.meta?.modem_db_path != null &&
-                    detectedTraceDB !== progress.meta?.modem_db_path
-                ) {
-                    detectedTraceDB = progress.meta?.modem_db_path;
-                    logger.info(`Using trace DB ${detectedTraceDB}`);
+                    detectedTraceDB = detectTraceDB(progress, detectedTraceDB);
+                    dispatch(setDetectingTraceDb(false));
                 }
 
                 progress.data_offsets